萬事從寬이면 其福自厚니라
해석:
If you follow generosity in everything, your blessings will grow on their own.

If generosity is followed in everything, the blessing naturally thickens

In other words, people should help others, but helping others is the principle of love and the best virtue of humans. It is like an easy principle that if you treat a person generously and warmly, the other person will follow and respect him, and if you are preparing to go a certain path, especially if you want to go up to a position that requires a lot of strength and support, it will be difficult to achieve your will unless you act and practice it.