忍時之忿이면 免百日憂니라
해석:
Enduring a once-in-a-lifetime anger saves you a hundred days of anxiety.

If you endure the anger at one time, you can avoid 100 days of anxiety, which means that it is difficult to endure, and if you endure it, it is as beneficial. Through the two kings of the Joseon Dynasty, we can see how much difference the results of patience differ.

King Sejong never got angry and blamed Choi Man-ri, a minor scholar of Jiphyeonjeon, who persistently opposed the creation of Hunminjeongeum. Rather, it is said that Sejong not only opposed to his will as a king, but also temporarily imprisoned him because he could not leave him in the way of creating Jeongum, and then visited him in prison at night to persuade him of his loyalty.

The rumor of Choi Man-ri's appeal against the creation of Jeongum served as an opportunity to encourage Jeongum creation to be more faithful and perfect. King Sejong's once anger suppressed the reign of the nation's greatest holy army.